pyproject.toml
[tool.poetry]
name = "event-manager"
version = "0.1.0"
description = "REST API for an app to manage events."
authors = ["volodymyr-v-konoval <volodymyr.v.konoval@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2"
djangorestframework = "^3.16.0"
drf-spectacular = "^0.28.0"
django-filter = "^25.1"
dj-database-url = "^2.3.0"
dotenv = "^0.9.9"
psycopg2-binary = "^2.9.10"
celery = "^5.5.2"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
